+++ index.js
@@ -349,7 +349,10 @@ module.exports = async function(options) {
       }
     }
     function test(name) {
-      if (!validSteps.includes(name)) {
+      // Projects that have different theme modules activated at different times
+      // are a frequent source of false positives for this warning, so ignore
+      // seemingly unused modules with "theme" in the name
+      if (!validSteps.includes(name) || (!name.match(/theme/))) {
         if (name.match(/^apostrophe-/)) {
           warn('namespace-apostrophe-modules', `You have a ${self.localModules}/${name} folder. You are probably trying to configure an official Apostrophe module, but those are namespaced now. Your directory should be renamed ${self.localModules}/${name.replace(/^apostrophe-/, '@apostrophecms/')}\n\nIf you get this warning for your own, original module, do not use the apostrophe- prefix. It is reserved.`);
         } else {