+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## UNRELEASED
+
+### Changes
+
+- Don't generate a warning about an unused module if it has "theme" in the name. This is a common pattern for modules that are not activated in all instances of the same codebase.
+
 ## 3.0.0-beta.2 - 2021-05-21
 
 ### **Breaks**